Failed eval #1, made minor changes. TODO:
changes:
Allocate memory for my struct on the stack instead of dynamic allocation
on the heap. Modified my build_struct() function and all other function
parameters in main() accordingly.
Removed a small error in my Makefile.
Added a previously missing newline to the
"No such file or directory" error in the child processes.

TODO:
Handle the case of an empty string "" or a single space " " being passed
as second or third argument. Currently it does not produce an error
message, even though it should.
???:
do something with WEXITSTATUS?
Handle cases like " cat "??????